feat(chat page - mobile) review and fix chat page.
